% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal.R
\name{bal}
\alias{bal}
\title{Construct Covariate Balance Statistics for Models with Multivariate Exposure}
\usage{
bal(
  model_list,
  D,
  C,
  all_uni = TRUE,
  common = FALSE,
  trim_w = FALSE,
  trim_quantile = 0.99
)
}
\arguments{
\item{model_list}{character string identifying which methods to use when
constructing weights. See details for a list of appropriate models}

\item{D}{numeric matrix of dimension \eqn{n} by \eqn{m} designating values of the exposures}

\item{C}{either a list of numeric matrices of length \eqn{m} of dimension
\eqn{n} by \eqn{p_j} designating values of the confounders for each exposure
value or if \code{common} is TRUE a single matrix of of dimension \eqn{n} by
\eqn{p} that represents common confounders for all exposures.}

\item{all_uni}{logical indicator. If TRUE then all univariate models specified
in model_list will be estimated for each exposure. If FALSE will only estimate weights
for the first exposure}

\item{common}{logical indicator for whether C is a single matrix of common
confounders for all exposures. default is FALSE meaning C must be specified
as list of confounders of length \eqn{m}.}

\item{trim_w}{logical indicator for whether to trim weights for mvGPS method. default is FALSE}

\item{trim_quantile}{numeric scalar used to specify the upper quantile to
trim weights if applicable. default is 0.99}
}
\description{
Construct Covariate Balance Statistics for Models with Multivariate Exposure
}
\details{
When using propensity score methods for causal inference it is crucial to
check the balancing property of the covariates and exposure(s).
}
