% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGPS.R
\name{D_C_check}
\alias{D_C_check}
\title{Internal function for formatting and checking specification of exposures and
confounders}
\usage{
D_C_check(D, C, common)
}
\arguments{
\item{D}{numeric matrix of dimension \eqn{n} by \eqn{m} designating values of the exposures}

\item{C}{either a list of numeric matrices of length \eqn{m} of dimension 
\eqn{n} by \eqn{p_j} designating values of the confounders for each exposure 
value or if \code{common} is TRUE a single matrix of of dimension \eqn{n} by
\eqn{p} that represents common confounders for all exposures.}

\item{common}{logical indicator for whether C is a single matrix of common
confounders for all exposures. default is FALSE meaning C must be specified
as list of confounders of length \eqn{m}.}
}
\description{
Internal function for formatting and checking specification of exposures and
confounders
}
\keyword{internal}
