% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_D.R
\name{gen_D}
\alias{gen_D}
\title{Generate Random Multivariate Exposure}
\usage{
gen_D(
  method,
  n,
  rho_cond,
  s_d1_cond,
  s_d2_cond,
  num_covs,
  C_mu,
  C_cov,
  C_var,
  d1_beta,
  d2_beta,
  seed = NULL
)
}
\arguments{
\item{method}{character value identifying which method so use when generating exposure. for bivariate model \code{uni_cond} is fastest}

\item{n}{integer value total number of units}

\item{rho_cond}{scalar value identifying conditional correlation of exposures given covariates between [0, 1]}

\item{s_d1_cond}{scalar value for conditional standard deviation of \code{D1}}

\item{s_d2_cond}{scalar value for conditional standard deviation of \code{D2}}

\item{num_covs}{integer value determining number of covariates to generate}

\item{C_mu}{numeric vector of mean values for covariates. must be same length as \code{num_covs}}

\item{C_cov}{scalar value representing constant correlation between covariates}

\item{C_var}{scalar value indicating constant variance of covariates}

\item{d1_beta}{numeric vector of length \code{num_covs} defining the mean of \code{D1} with respect to the covariates}

\item{d2_beta}{numeric vector of length \code{num_covs} defining the mean of \code{D2} with respect to the covariates}

\item{seed}{integer value setting the seed of random generator to produce repeatable results. set to NULL by default}
}
\description{
This function allows the user to generate exposure from a multivariate normal distribution confounded by a set of variables \code{C}.
}
