% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull_sample.R
\name{hull_sample}
\alias{hull_sample}
\title{Sample Points Along a Convex Hull}
\usage{
hull_sample(X, num_grid_pts = 500)
}
\arguments{
\item{X}{numeric matrix of n by p dimensions. Each row corresponds to a point in p-dimensional space.}

\item{num_grid_pts}{integer scalar denoting the number of parameters to
search for over the convex hull. Default is 500.}
}
\description{
Sample Points Along a Convex Hull
}
