% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGPS.fit.R
\name{mvGPS.fit}
\alias{mvGPS.fit}
\title{Fit Weighted Linear Regression for Dose-Response}
\usage{
mvGPS.fit(formula, W, data = sys.frame(sys.parent()))
}
\arguments{
\item{formula}{object of class \code{formula} which symbolically describes
the dose-response with the outcome on one side and a function of the exposures on
the right side.}

\item{W}{list of weights to use in the dose-response regression}

\item{data}{data.frame that contains the outcome and covariates for
dose-response. If none is provided then assumes objects are in user's environment.}
}
\description{
Fitting the dose-response model for causal inference model using weights
generated from various propensity score models. Can use either the
\code{\link[mvGPS]{mvGPS}} function to generate weights or the
\code{\link[mvGPS]{bal}} function to generate entire list of weights from
different propensity models.
}
\details{
Will describe details here
}
