% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGPS_out_fit.R
\name{mvGPS_out_fit}
\alias{mvGPS_out_fit}
\title{Fit linear outcome regression using potential weights}
\usage{
mvGPS_out_fit(
  Y,
  W,
  D,
  C,
  num_grid_pts = 500,
  poly_degree = NULL,
  D_interact = FALSE
)
}
\arguments{
\item{Y}{numeric vector specifying the outcome of interest}

\item{W}{list of weights to use in the outcome}

\item{D}{numeric matrix of dimension \code{n} by \code{m} designating values of the exposure}

\item{C}{list of numeric matrices of length \code{m} of dimension \code{n} by \code{k} designating values of the confounders for each exposure value}

\item{num_grid_pts}{integer scalar denoting the number of parameters to
search for over the convex hull. Default is 500.}

\item{poly_degree}{integer scalar specifying the polynomial degree which will
be applied to each exposure and included in the outcome modeling. Default is NULL
signifying that only first order terms are included}

\item{D_interact}{logical indicator denoting whether to include all pairwise
interactions of univariate exposures when modeling}
}
\description{
Fit linear outcome regression using potential weights
}
\details{
Availables models include "mvGPS", "entropy", "CBPS", "PS"
}
