% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGPSsim_results.R
\name{mvGPSsim_results}
\alias{mvGPSsim_results}
\title{Construct Summary Statistics for Simulation Models}
\usage{
mvGPSsim_results(model_list, all_uni = TRUE, D, C, alpha)
}
\arguments{
\item{model_list}{character string identifying which methods to use when constructing weights. See details for a list of appropriate models}

\item{all_uni}{logical indicator. If TRUE then all univariate models specified
in model_list will be estimated for each exposure. If FALSE will only estimate weights
for the first exposure}

\item{D}{numeric matrix of dimension \code{n} by \code{2} designating values of the exposure}

\item{C}{numeric matrix of dimension \code{n} by \code{k} designating values of the confounders}

\item{alpha}{numeric vector of length \code{k+2+1} used for constructing the mean of the outcome \code{Y}.
The first term represents the intercept, the next \code{k} terms denote the effect of the confounders,
and the final \code{2} terms represent the true treatment effect
where \code{k} represents the number of confounders}
}
\description{
Construct Summary Statistics for Simulation Models
}
\details{
Availables models include "mvGPS", "entropy", "CBPS", "PS"
}
