% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull_sample.R
\name{X_check}
\alias{X_check}
\title{Checking that the exposure matrix is properly specified}
\usage{
X_check(X)
}
\arguments{
\item{X}{numeric matrix of n by m dimensions. Each row corresponds to a point in m-dimensional space.}
}
\description{
Checking that the exposure matrix is properly specified
}
\keyword{internal}
