% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGPS.R
\name{mvGPS}
\alias{mvGPS}
\title{Multivariate Generalized Propensity Score (mvGPS)}
\usage{
mvGPS(formula, data = sys.frame(sys.parent()))
}
\arguments{
\item{formula}{an object of class formula that describes the generalized
propensity score model. The outcome must be contain at least two exposures}

\item{data}{data.frame that contains variables used in formula}
}
\description{
Construct weights using propensity scores for multivariate continuous exposure
by assuming joint normal marginal and conditional densities. See details for
description of weights and additional methodology details.
}
