% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGPS.R
\name{mvGPS}
\alias{mvGPS}
\title{Multivariate Generalized Propensity Score (mvGPS)}
\usage{
mvGPS(D, C, trim_w = FALSE, trim_quantile = 0.99)
}
\arguments{
\item{D}{numeric matrix of dimension \code{n} by \code{m} designating values of the exposure}

\item{C}{list of numeric matrices of length \code{m} of dimension \code{n} by \code{k} designating values of the confounders for each exposure value}

\item{trim_w}{logical indicator for whether to trim weights. default is FALSE}

\item{trim_quantile}{numeric scalar used to specify the upper quantile to
trim weights if applicabile. default is 0.99}
}
\description{
Construct weights using propensity scores for multivariate continuous exposure
by assuming joint normal marginal and conditional densities. See details for
description of weights and additional methodology details.
}
