% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGPSsim_bal.R
\name{mvGPSsim_bal}
\alias{mvGPSsim_bal}
\title{Construct Covariate Balance Statistics for Simulation Models}
\usage{
mvGPSsim_bal(model_list, D, C, all_uni = TRUE, trim_w = FALSE)
}
\arguments{
\item{model_list}{character string identifying which methods to use when constructing weights. See details for a list of appropriate models}

\item{D}{numeric matrix of dimension \code{n} by \code{2} designating values of the exposure}

\item{C}{numeric matrix of dimension \code{n} by \code{k} designating values of the confounders}

\item{all_uni}{logical indicator. If TRUE then all univariate models specified
in model_list will be estimated for each exposure. If FALSE will only estimate weights
for the first exposure}

\item{trim_w}{logical indicator for whether to trim weights for mvGPS method. default is FALSE}
}
\description{
Construct Covariate Balance Statistics for Simulation Models
}
