% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvGPSsim_results.R
\name{mvGPSsim_results}
\alias{mvGPSsim_results}
\title{Construct Summary Statistics for Simulation Models}
\usage{
mvGPSsim_results(
  W,
  D,
  C,
  alpha,
  sd_Y,
  num_grid_pts = 500,
  perf_metrics = c("MSE", "bias"),
  poly_degree = NULL,
  D_interact = FALSE
)
}
\arguments{
\item{W}{list of weights to use in the outcome}

\item{D}{numeric matrix of dimension \code{n} by \code{2} designating values of the exposure}

\item{C}{numeric matrix of dimension \code{n} by \code{k} designating values of the confounders}

\item{alpha}{numeric vector of length \code{k+2+1} used for constructing the mean of the outcome \code{Y}.
The first term represents the intercept, the next \code{k} terms denote the effect of the confounders,
and the final \code{2} terms represent the true treatment effect
where \code{k} represents the number of confounders}

\item{sd_Y}{numeric scalar designating the standard deviation to assume when
generating the outcome. Error term is assummed to be normally distributed.}

\item{num_grid_pts}{integer scalar denoting the number of parameters to
search for over the convex hull. Default is 500.}

\item{perf_metrics}{character vector of which performance metrics to return.
Default is to return "MSE" and "bias". See details for other options.}

\item{poly_degree}{integer scalar specifying the polynomial degree which will
be applied to each exposure and included in the outcome modeling. Default is NULL
signifying that only first order terms are included}

\item{D_interact}{logical indicator denoting whether to include all pairwise
interactions of univariate exposures when modeling}
}
\description{
Construct Summary Statistics for Simulation Models
}
\details{
Availables models include "mvGPS", "entropy", "CBPS", "PS"
}
